fix: use aarch64 suffix instead of arm64
On Linux, arm64 architecture is commonly referred to as aarch64. The
`uname -m` command returns the latter. As a result, the command
suggested in the README to install dockerize doesn't work on this
architecture.

This commit fixes this by renaming the release artifact from
"linux-arm64" to "linux-aarch64". Thus, "$(uname -s)-$(uname -m)" will
return the expected artifact name.

See also: https://stackoverflow.com/a/45125525